+++ tw-tkms-starter/src/main/java/com/transferwise/kafka/tkms/ITkmsPaceMaker.java
@@ -1,12 +1,15 @@
 package com.transferwise.kafka.tkms;
 
+import com.transferwise.kafka.tkms.api.TkmsShardPartition;
 import java.time.Duration;
 
 public interface ITkmsPaceMaker {
 
   void doSmallPause(int shard);
-
-  void pauseOnError(int shard);
-
+  
   Duration getLongWaitTime(int shard);
+
+  Duration getPollingPause(TkmsShardPartition tkmsShardPartition, int pollingBatchSize, int polledMessagesCount);
+
+  Duration getPollingPauseOnError(TkmsShardPartition tkmsShardPartition);
 }