+++ tw-tkms-starter/src/test/java/com/transferwise/kafka/tkms/EarliestMessageTrackingIntTest.java
@@ -10,10 +10,12 @@ import com.transferwise.kafka.tkms.api.TkmsShardPartition;
 import com.transferwise.kafka.tkms.config.TkmsProperties;
 import com.transferwise.kafka.tkms.dao.ITkmsDao;
 import com.transferwise.kafka.tkms.metrics.ITkmsMetricsTemplate;
+import com.transferwise.kafka.tkms.metrics.TkmsMetricsTemplate;
 import com.transferwise.kafka.tkms.test.BaseIntTest;
 import io.micrometer.core.instrument.Gauge;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
+import java.util.Objects;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -42,7 +44,7 @@ public class EarliestMessageTrackingIntTest extends BaseIntTest {
     TkmsClockHolder.setClock(clock);
 
     Gauge earliestMessageIdGauge = await()
-        .until(() -> meterRegistry.find("tw.tkms.dao.earliest.message.id").tags("shard", "0", "partition", "0").gauge(), (g) -> g != null);
+        .until(() -> meterRegistry.find(TkmsMetricsTemplate.DAO_EARLIEST_MESSAGE_ID).tags("shard", "0", "partition", "0").gauge(), Objects::nonNull);
     assertThat(earliestMessageIdGauge.value()).isEqualTo(-1);
 
     clock.tick(Duration.ofSeconds(5));
@@ -52,7 +54,7 @@ public class EarliestMessageTrackingIntTest extends BaseIntTest {
 
     clock.tick(Duration.ofSeconds(6));
     sendMessageAndWaitForArrival(2);
-    assertThat(earliestMessageIdGauge.value()).as("First 10s period has passed, we should have a id now").isGreaterThanOrEqualTo(0);
+    assertThat(earliestMessageIdGauge.value()).as("First 10s period has passed, we should have a id now").isNotNegative();
 
     double previousValue = earliestMessageIdGauge.value();
     clock.tick(Duration.ofSeconds(6));
@@ -78,7 +80,7 @@ public class EarliestMessageTrackingIntTest extends BaseIntTest {
     transactionsHelper.withTransaction().run(() ->
         tkms.sendMessage(new TkmsMessage().setTopic(testTopic).setValue("Hello World!".getBytes(StandardCharsets.UTF_8)))
     );
-    
+
     await().until(() -> tkmsSentMessagesCollector.getSentMessages(testTopic).size() == cnt);
   }
 }