+++ tw-tkms-starter/src/test/java/com/transferwise/kafka/tkms/MonitoringIntTest.java
@@ -2,6 +2,7 @@ package com.transferwise.kafka.tkms;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.transferwise.kafka.tkms.metrics.TkmsMetricsTemplate;
 import com.transferwise.kafka.tkms.test.BaseIntTest;
 import io.micrometer.core.instrument.Gauge;
 import org.awaitility.Awaitility;
@@ -12,12 +13,12 @@ class MonitoringIntTest extends BaseIntTest {
   @Test
   void testThatMonitoringMetricsArePresent() {
     Awaitility.await().until(() -> {
-      Gauge gauge = meterRegistry.find("tw.tkms.dao.approximate.messages.count").tags("shard", "0", "partition", "0").gauge();
+      Gauge gauge = meterRegistry.find(TkmsMetricsTemplate.DAO_APPROXIMATE_MESSAGES_COUNT).tags("shard", "0", "partition", "0").gauge();
       return gauge != null && gauge.value() >= 0;
     });
 
     Awaitility.await().until(() -> {
-      Gauge gauge = meterRegistry.find("tw.tkms.dao.approximate.messages.count").tags("shard", "1", "partition", "0").gauge();
+      Gauge gauge = meterRegistry.find(TkmsMetricsTemplate.DAO_APPROXIMATE_MESSAGES_COUNT).tags("shard", "1", "partition", "0").gauge();
       return gauge != null && gauge.value() >= 0;
     });
   }
@@ -25,12 +26,12 @@ class MonitoringIntTest extends BaseIntTest {
   @Test
   void testThatTableStatsMetricsArePresent() {
     Awaitility.await().until(() -> {
-      Gauge gauge = meterRegistry.find("tw.tkms.dao.rows.in.table.stats").tags("shard", "1", "partition", "0").gauge();
+      Gauge gauge = meterRegistry.find(TkmsMetricsTemplate.DAO_ROWS_IN_TABLE_STATS).tags("shard", "1", "partition", "0").gauge();
       return gauge != null && gauge.value() == 1_000_000;
     });
 
     Awaitility.await().until(() -> {
-      Gauge gauge = meterRegistry.find("tw.tkms.dao.rows.in.index.stats").tags("shard", "1", "partition", "0").gauge();
+      Gauge gauge = meterRegistry.find(TkmsMetricsTemplate.DAO_ROWS_IN_INDEX_STATS).tags("shard", "1", "partition", "0").gauge();
       return gauge != null && gauge.value() == 1_000_000;
     });
   }
@@ -38,11 +39,11 @@ class MonitoringIntTest extends BaseIntTest {
   @Test
   void earliestMessageIdIsRegistered() {
     Awaitility.await().until(() -> {
-      Gauge gauge = meterRegistry.find("tw.tkms.dao.earliest.message.id").tags("shard", "1", "partition", "0").gauge();
+      Gauge gauge = meterRegistry.find(TkmsMetricsTemplate.DAO_EARLIEST_MESSAGE_ID).tags("shard", "1", "partition", "0").gauge();
       return gauge != null;
     });
 
-    assertThat(meterRegistry.find("tw.tkms.dao.earliest.message.id").tags("shard", "0", "partition", "0").gauge())
+    assertThat(meterRegistry.find(TkmsMetricsTemplate.DAO_EARLIEST_MESSAGE_ID).tags("shard", "0", "partition", "0").gauge())
         .as("Earliest message id tracking is not enabled for shard 0.").isNull();
   }
 }