+++ tw-tkms-starter/src/test/java/com/transferwise/kafka/tkms/PostgresMonitoringIntTest.java
@@ -1,5 +1,6 @@
 package com.transferwise.kafka.tkms;
 
+import com.transferwise.kafka.tkms.metrics.TkmsMetricsTemplate;
 import io.micrometer.core.instrument.Gauge;
 import org.awaitility.Awaitility;
 import org.junit.jupiter.api.Test;
@@ -11,7 +12,7 @@ public class PostgresMonitoringIntTest extends MonitoringIntTest {
   @Test
   void testThatTableStatsMetricsArePresent() {
     Awaitility.await().until(() -> {
-      Gauge gauge = meterRegistry.find("tw.tkms.dao.rows.in.table.stats").tags("shard", "1", "partition", "0").gauge();
+      Gauge gauge = meterRegistry.find(TkmsMetricsTemplate.DAO_ROWS_IN_TABLE_STATS).tags("shard", "1", "partition", "0").gauge();
       return gauge != null && gauge.value() == 1_000_000;
     });
   }