+++ tw-tkms-starter/src/test/java/com/transferwise/kafka/tkms/test/BaseIntTest.java
@@ -23,7 +23,7 @@ public class BaseIntTest {
 
   @Autowired
   protected IMeterCache meterCache;
-  
+
   @Autowired
   protected JdbcTemplate jdbcTemplate;