+++ tw-tkms-starter/src/test/java/db/migration/earliestmessage/V2__Init.java
@@ -18,7 +18,7 @@ public class V2__Init extends BaseJavaMigration {
           stmt.execute("CREATE TABLE " + tableName + " (\n"
               + "  id BIGSERIAL PRIMARY KEY,\n"
               + "  message BYTEA NOT NULL\n"
-              + ") WITH (autovacuum_analyze_threshold=1000000000, autovacuum_vacuum_threshold=100000, toast_tuple_target=8160) ");
+              + ") WITH (autovacuum_analyze_threshold=1000000000, toast_tuple_target=8160) ");
           log.info("Create table `" + tableName + "'.");
 
           stmt.executeUpdate("ALTER TABLE " + tableName + " ALTER COLUMN id SET (n_distinct=1000000);");