+++ CHANGELOG.md
@@ -5,11 +5,60 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.0] - 2023-01-31
+
+### Changed
+
+#### Messages polling interval is now more dynamic.
+
+We are pausing for (pollingInterval * (batchSize - polledRecords) / batchSize).
+E.g.
+
+- when we poll a full batch, we will not wait at all and start the next cycle immediately.
+- when we poll zero records, we will wait the full polling interval.
+- when we get half-full batch, we will be waiting 50% from polling interval.
+
+This is most useful, for Postgres databases, where workloads are somewhat HTAP, or autovacuum is not snappy enough to clean dead tuples, which every
+poll has to visit.
+
+The default logic can be overridden by custom implementation of `ITkmsPaceMaker`.
+
+`tw_tkms_proxy_cycle` timer does not include those pauses anymore.
+`tw_tkms_proxy_cycle_pause` timer is added to specifically measure those pauses.
+
+#### Index hints for Postgres queries.
+
+It turned out, that the `n_distinct` trick we suggested in setup, did not actually apply in all scenarios. We still had an incident where delete
+queries started to do full sequential scans.
+
+Now we will be relying on `pg_hint_plan` extension being available. Fortunately RDS is supporting it.
+
+Some initialization validation routines were added checking if index hints actually do apply.
+
+Removed the recommendation of setting `autovacuum_vacuum_threshold=100000` for tkms tables. Auto vacuum has its own, database side, `naptime` setting
+to prevent it running in a tight loop.
+
+#### Configurable delete queries batch sizes.
+
+Services can now configure delete queries batch sizes. This can be useful, when the database is still trying to do sequential scans let's say with
+1024 parameters, but would not do it with 256.
+
+This can be done via the `deleteBatchSizes` property.
+
+#### Initialization validations
+
+Added more initialization validations around database performance.
+
+#### MariaDb fixed stats
+
+The stats value recommendations for `stat_value` and `n_rows` was 
+
 ## [0.20.0] - 2023-01-23
 
 ### Changed
 
-* `TransactionalKafkaMessageSender` is now throwing an error when messages are tried to be registered without active transactions.
+* Improved graceful shutdown.
+  `TransactionalKafkaMessageSender` is now throwing an error when messages are tried to be registered without active transactions.
   It helps to easily detect issues, where database changes from business logic and Kafka messages sending are happening in separate transactions.
   If there is no active transaction, the overhead from `Tkms` does not make any sense.
   The active transaction check can be disabled by setting `requireTransactionOnMessagesRegistering` property to `false`.